requirements.txt
shopify-python-api
python-dotenv